requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
folium==0.14.0
streamlit-folium==0.15.0
numpy==1.24.3
Pillow==10.0.1
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
streamlit-elements==0.1.0
streamlit-ace==0.1.1
